refactor: split dynamics into submodules
